services/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic-settings==2.5.2
python-multipart==0.0.12
asyncpg==0.29.0
httpx==0.27.2
pyjwt==2.9.0
bcrypt==4.2.0
prometheus-client==0.20.0
